api/requirements.txt
alembic==1.8.1
async-timeout==4.0.2
bcrypt==4.0.0
cachelib==0.9.0
click==8.1.3
Deprecated==1.2.13
Flask==2.2.2
Flask-Bcrypt==1.0.1
Flask-Cors==3.0.10
Flask-JWT-Extended==4.4.4
Flask-Migrate==3.1.0
Flask-Session==0.4.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.3
gunicorn==20.1.0
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.3
MarkupSafe==2.1.1
packaging==21.3
psycopg2-binary==2.9.6
PyJWT==2.5.0
pyparsing==3.0.9
python-dotenv==0.21.0
redis==4.3.4
six==1.16.0
SQLAlchemy==1.4.41
Werkzeug==2.2.2
wrapt==1.14.1
zipp==3.8.1
